move LandingPage and UserMenu into a folder "Main"
